feat(core): move webdriver utils to new core package
